+++ vk_descriptor.h
@@ -24,7 +24,7 @@ void vk_free_descriptor_set(VkDevice device, VkDescriptorPool descriptor_pool, V
 
 void vk_update_descriptor_sets(VkDevice device, uint32_t descriptor_write_count, const VkWriteDescriptorSet *writes);
 
-VkDescriptorBufferInfo vk_descriptor_buffer_info(VkBuffer buffer, uint32_t size);
+VkDescriptorBufferInfo vk_descriptor_buffer_info(VkBuffer buffer, uint64_t size);
 VkDescriptorImageInfo vk_descriptor_image_info(VkSampler sampler, VkImageView image_view, VkImageLayout image_layout);
 
 VkWriteDescriptorSet vk_write_descriptor_set(VkDescriptorSet descriptor_set, uint32_t binding, VkDescriptorType descriptor_type, const VkDescriptorImageInfo *image_info, const VkDescriptorBufferInfo *buffer_info);
